+++ alacritty/Cargo.toml
@@ -13,6 +13,7 @@ rust-version = "1.70.0"
 path = "../alacritty_terminal"
 version = "0.20.0-dev"
 default-features = false
+features = ["serde"]
 
 [dependencies.alacritty_config_derive]
 path = "../alacritty_config_derive"
