+++ alacritty_terminal/src/tty/windows/conpty.rs
@@ -17,11 +17,11 @@ use windows_sys::Win32::System::Threading::{
     STARTF_USESTDHANDLES, STARTUPINFOEXW, STARTUPINFOW,
 };
 
-use crate::config::PtyConfig;
 use crate::event::{OnResize, WindowSize};
 use crate::tty::windows::blocking::{UnblockedReader, UnblockedWriter};
 use crate::tty::windows::child::ChildExitWatcher;
 use crate::tty::windows::{cmdline, win32_string, Pty};
+use crate::tty::PtyConfig;
 
 const PIPE_CAPACITY: usize = crate::event_loop::READ_BUFFER_SIZE;
