+++ alacritty_terminal/src/vi_mode.rs
@@ -1,6 +1,7 @@
 use std::cmp::min;
 
-use alacritty_config_derive::ConfigDeserialize;
+#[cfg(feature = "serde")]
+use serde::{Deserialize, Serialize};
 
 use crate::event::EventListener;
 use crate::grid::{Dimensions, GridCell};
@@ -9,7 +10,8 @@ use crate::term::cell::Flags;
 use crate::term::Term;
 
 /// Possible vi mode motion movements.
-#[derive(ConfigDeserialize, Debug, Copy, Clone, PartialEq, Eq)]
+#[derive(Debug, Copy, Clone, PartialEq, Eq)]
+#[cfg_attr(feature = "serde", derive(Serialize, Deserialize))]
 pub enum ViMotion {
     /// Move up.
     Up,
@@ -380,15 +382,14 @@ mod tests {
     use super::*;
 
     use crate::ansi::Handler;
-    use crate::config::Config;
     use crate::event::VoidListener;
     use crate::index::{Column, Line};
     use crate::term::test::TermSize;
-    use crate::term::Term;
+    use crate::term::{Term, TermOptions};
 
     fn term() -> Term<VoidListener> {
         let size = TermSize::new(20, 20);
-        Term::new(&Config::default(), &size, VoidListener)
+        Term::new(TermOptions::default(), &size, VoidListener)
     }
 
     #[test]
